fix!: clock source miss-match between Rust and Go (#1742)
## Why this should be merged

coarsetime was introduced to avoid syscall overhead from `std::time`,
but @demosdemon confirmed (and we verified empirically) that std::time
uses `clock_gettime(CLOCK_MONOTONIC)` via vDSO on modern Linux which is
resolving in userspace with no kernel entry. The two are equivalent in
cost.

Removing coarsetime fixes the clock source mismatch between Rust
`CLOCK_MONOTONIC_COARSE` and Go `CLOCK_MONOTONIC` that was causing the
Firewood dashboard to show Rust commit time higher than Go commit time,
which is physically impossible since Go wraps Rust synchronously via
CGO.

## How this works

Replaces all `coarsetime::Instant` / `coarsetime::Duration` usages
with`std::time::Instant` / `std::time::Duration`

## How this was tested

Full methodology in #1720.

## Breaking Changes

- [ ] firewood
- [ ] firewood-storage
- [x] firewood-ffi (C api) — `commit_proposal` token closure parameter
type changes from `coarsetime::Duration` to `std::time::Duration`
- [ ] firewood-go (Go api)
- [ ] fwdctl

+++ storage/src/nodestore/persist.rs
@@ -31,8 +31,8 @@ use bumpalo::Bump;
 use std::iter::FusedIterator;
 
 use crate::linear::FileIoError;
-use coarsetime::Instant;
 use firewood_metrics::firewood_increment;
+use std::time::Instant;
 
 use crate::{MaybePersistedNode, NodeReader, WritableStorage};
 
@@ -191,7 +191,7 @@ impl<S: WritableStorage> NodeStore<Committed, S> {
 
         self.process_unpersisted_nodes(&mut bump, &mut node_allocator, super::INITIAL_BUMP_SIZE)?;
 
-        let flush_time = flush_start.elapsed().as_millis();
+        let flush_time = flush_start.elapsed().as_millis() as u64;
         firewood_increment!(crate::registry::FLUSH_NODES, flush_time);
 
         Ok(())
